Be specific about reading a byte
Previously the code was relying on "readStr" to do double-duty as
reading a single byte as a byte and reading an encoded string. This is
dangerous as its harder to dectect errors in the protocol. Furthermore,
it decoded the content based on the stdout oncoding which is related to
the terminal encoding rather than the underlying protocol. The
documentation I read (https://help.mikrotik.com/docs/display/ROS/API)
did *not* specify a character encoding for the "attribute words".

I'm going to assume UTF-8 until I get stuff that breaks.

This means that I can fix the readStr function to error out on decoding
errors and always return a valid string.